Fix reaper not working after cluster is recreated. (#268)
* Fix reaper not working after cluster is recreated.
Because of all nodes changed their IPs reaper can't connect to any host to rediscover the cluster: https://github.com/thelastpickle/cassandra-reaper/issues/993#issuecomment-812362147
Unfortunately Reaper doesn't provide away to reliably detect that reaper doesn't work well so we have to check the symptoms.
The symptoms are a failing GET /cluster/<cluster-name> request with timeout error. The timeout happens because reaper tries to connect to Cassandra hosts but can't because it tries to connect with the wrong IPs. So when it's happen we also check if the cluster exists in the list of clusters (GET /cluster). If that succeeds we assume that it's a broken reaper. In that case we have re init reaper by re-adding the cluster.
After re-adding we also need to restart reaper pods since there are some leftover info in the memory about the old IPs and reaper will continue trying to connect to those nodes.